* Day 11 Part 1

+++ AdventOfCode2020/Day8.cpp
@@ -125,7 +125,7 @@ int Day8::PartTwo()
 			ExecuteNext();
 
 			// SUCCESS! If the program has terminated successfully, return the accumulator value
-			if (_instructionIdx == _instructions.size())
+			if (_instructionIdx == static_cast<int>(_instructions.size()))
 				return _accumulator;
 		}
